Reconstruct the project, move corresponding files to component, style, asset
Create Header, Body, Footer components and play around with Body css
