+++ react-cttc/src/component/App.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import Header from "./Header";
+import Body from "./Body";
+import Footer from "./Footer";
+import '../style/App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <Header/>
+      <Body/>
+      <Footer/>
+    </div>
+  );
+}
+
+export default App;
