+++ react-cttc/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import './style/index.css';
+import App from './component/App';
+import reportWebVitals from './tst/reportWebVitals';
 
 ReactDOM.render(
   <React.StrictMode>
