+++ react-cttc/src/style/App.css
@@ -1,0 +1,32 @@
+@import "./Header.css";
+@import "./Body.css";
+@import "./Footer.css";
+
+.App {
+  text-align: center;
+}
+
+.App-logo {
+  height: 10vmin;
+  pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin 5s infinite linear;
+  }
+}
+
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
+
+.CTTC {
+  font-size: 8vmin;
+  color: white;
+}
