+++ react-cttc/src/style/Footer.css
@@ -1,0 +1,5 @@
+.App-footer {
+    background-color: #282c34;
+    height: 10vmin;
+    display: flex;
+}
