+++ react-cttc/src/style/Header.css
@@ -1,0 +1,5 @@
+.App-header {
+    background-color: #282c34;
+    height: 10vmin;
+    display: flex;
+}
